Improvements in db updation.
Some checks had to be refined. Some cronjob specific changes were needed.

+++ cron_jobs/get_updated_dbs.sh
@@ -1,13 +1,17 @@
 #!/bin/bash -e
-
+export USER=plumgrid
 export PATH=@CMAKE_BINARY_DIR@/:/opt/pg/scripts:/opt/pg/scripts/gcloud/aurora/pipeline_scripts/:$PATH
 source gc_helpers.sh
 source pl_utils.sh
 source pl_settings.conf
 source aurora_infra_settings.conf
 source pl_help.sh
+source /home/${USER}/.aurora.conf
 
-TEMP=`getopt -o W: --long WORKSPACE: -n 'update_db.sh' -- "$@"`
+TEST_FAILURES_ACCEPTABLE=10
+TEST_FAILURES_ACCEPTABLE_UNSTABLE=42
+
+TEMP=`getopt -o W: --long WORKSPACE: -n 'get_updated_db.sh' -- "$@"`
 eval set -- "$TEMP"
 
 while true ; do
@@ -31,13 +35,12 @@ trap "cleanup ${BUILD_ID}" EXIT
 
 create_dir_or_remove_contents $WORKSPACE
 
-declare -a branches=('master' 'stable_4_0' 'stable_4_1');
-declare -a pipelines=('smoke' 'extended' 'unstable' 'omni' 'pgui-smoke');
+declare -a branches=('master' 'stable_4_0');
+declare -a pipelines=('smoke' 'extended' 'omni' 'unstable' 'pgui-smoke');
 
 GCLOUD_PATH=${WORKSPACE}/gcloud
-
-git clone ssh://${GERRIT_ID}@${GERRIT_IP}:${GERRIT_PORT}/gcloud.git ${GCLOUD_PATH}
-scp -p -P ${GERRIT_PORT} ${GERRIT_ID}@${GERRIT_IP}:hooks/commit-msg ${GCLOUD_PATH}/.git/hooks/
+git clone ssh://${gerritid}@${GERRIT_IP}:${GERRIT_PORT}/gcloud.git ${GCLOUD_PATH}
+scp -p -P ${GERRIT_PORT} ${gerritid}@${GERRIT_IP}:hooks/commit-msg ${GCLOUD_PATH}/.git/hooks/
 
 
 for BRANCH in ${branches[*]}; do
@@ -45,15 +48,18 @@ for BRANCH in ${branches[*]}; do
   for PIPELINE in ${pipelines[*]}; do
     tag="dbalps"
     project="alps"
-    acceptable_test_failures=10
+    acceptable_test_failures=$TEST_FAILURES_ACCEPTABLE
     if [[ "$PIPELINE" == "unstable" ]]; then
-      acceptable_test_failures=38
+      acceptable_test_failures=$TEST_FAILURES_ACCEPTABLE_UNSTABLE
     elif [[ "$PIPELINE" == "pgui-smoke" ]]; then
       project="pg_ui"
       tag="dbpgui"
     fi
     # this build id would be created, the variable BUILD_ID is for cleanup, since we do not have any information about the build.
     BUILD_ID=${tag}-bld-${BRANCH}
+    if [[ $BRANCH == "stable_4_0" ]]; then
+      BUILD_ID="${tag}-bld-stable-4-0"
+    fi
     PIPELINE_WORKSPACE=${WORKSPACE}/pl-${PIPELINE}-${BRANCH}
     aurora update_db -p ${project} -P ${PIPELINE} -b ${BRANCH} -W ${WORKSPACE} -u ${acceptable_test_failures} -t ${tag}
     # After the end of run pipelines, after the updation of db, the db is copied into the workspace of the pipelines in a dir dbs
@@ -62,11 +68,13 @@ for BRANCH in ${branches[*]}; do
       path_to_created_db=$(get_db $PIPELINE $BRANCH)
       pushd ${GCLOUD_PATH}
       git reset --hard origin/master
-      echo ${WORKSPACE}/${PIPELINE_WORKSPACE}/dbs
       echo ${GCLOUD_PATH}/aurora/pipeline_scripts/alps/dbs
       # Remove the previous db in the newly pulled repo.
-      rm ${GCLOUD_PATH}/aurora/pipeline_scripts/alps/dbs/${PIPELINE}-${BRANCH}.db
+      if [[ -e "${GCLOUD_PATH}/aurora/pipeline_scripts/alps/dbs/${PIPELINE}-${BRANCH}.db" ]]; then
+        rm ${GCLOUD_PATH}/aurora/pipeline_scripts/alps/dbs/${PIPELINE}-${BRANCH}.db
+      fi
       # Copy new db to the dbs directory in the newly pulled gcloud repo to commit.
+      echo "copying ${path_to_created_db} *** to *** ${GCLOUD_PATH}/aurora/pipeline_scripts/alps/dbs/"
       cp ${path_to_created_db} ${GCLOUD_PATH}/aurora/pipeline_scripts/alps/dbs/
       git status
       cat /dev/null > commit_msg_file
@@ -77,7 +85,7 @@ for BRANCH in ${branches[*]}; do
       cat commit_msg_file
 
       # Commit the file
-      git commit -s -o -F commit_msg_file aurora/pipeline_scripts/alps/${PIPELINE}-${BRANCH}.db
+      git commit -s -o -F commit_msg_file aurora/pipeline_scripts/alps/dbs/${PIPELINE}-${BRANCH}.db
       # Push the commit
       git push origin HEAD:refs/for/master
       rm commit_msg_file
@@ -86,6 +94,6 @@ for BRANCH in ${branches[*]}; do
     fi
   done
   # cleaning up the instances that were used for the previous db update
-  echo "cleaning up instances related to $BUILD_ID"
+  echo "cleaning up instances related to ${BUILD_ID}"
   cleanup $BUILD_ID
 done
